pyproject.toml
[project]
name = "sia-tp2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.4",
    "pillow>=11.1.0",
]
